requirements.txt
python-telegram-bot==20.6
opencv-python-headless
numpy
pillow
httpx
